xtc-prepare.sh should not rebuild every xtc run

+++ Briar/xtc-prepare.sh
@@ -2,9 +2,16 @@
 
 XAMARIN_DIR="${PWD}/xtc-staging"
 
-echo "INFO: creating the ./xtc-staging directory"
-rm -rf "${XAMARIN_DIR}"
-mkdir -p "${XAMARIN_DIR}"
+
+if [ "$1" = "-" ]; then
+    echo "INFO: passed a '$1' so we will try to reuse the .ipa"
+    echo "INFO: deleting the ./xtc-staging/features directory"
+    rm -rf "${XAMARIN_DIR}/features"
+else
+    echo "INFO: creating the ./xtc-staging directory"
+    rm -rf "${XAMARIN_DIR}"
+    mkdir -p "${XAMARIN_DIR}"
+fi
 
 echo "INFO: copying features over to ${XAMARIN_DIR}"
 cp -r features "${XAMARIN_DIR}/"
@@ -18,45 +25,49 @@ BRIAR_INSTALL=`bundle show briar`
 BRIAR_STEPS="${BRIAR_INSTALL}/features/step_definitions"
 cp -r "${BRIAR_STEPS}" "${XAMARIN_DIR}/features/step_definitions/briar"
 
-WORKSPACE="../briar-ios-example.xcworkspace"
-SCHEME="Briar-cal"
-TARGET_NAME="Briar-cal"
-CONFIG="Debug"
+if [ "$1" = "-" ]; then
+    echo "INFO: will not rebuild the .ipa"
+else
+    WORKSPACE="../briar-ios-example.xcworkspace"
+    SCHEME="Briar-cal"
+    TARGET_NAME="Briar-cal"
+    CONFIG="Debug"
 
-CAL_DISTRO_DIR="build/calabash"
-ARCHIVE_BUNDLE="${CAL_DISTRO_DIR}/briar-cal.xcarchive"
-APP_BUNDLE_PATH="${ARCHIVE_BUNDLE}/Products/Applications/Briar-cal.app"
-IPA_PATH="${CAL_DISTRO_DIR}/${TARGET_NAME}.ipa"
+    CAL_DISTRO_DIR="build/calabash"
+    ARCHIVE_BUNDLE="${CAL_DISTRO_DIR}/briar-cal.xcarchive"
+    APP_BUNDLE_PATH="${ARCHIVE_BUNDLE}/Products/Applications/Briar-cal.app"
+    IPA_PATH="${CAL_DISTRO_DIR}/${TARGET_NAME}.ipa"
 
-rm -rf "${CAL_DISTRO_DIR}"
-mkdir -p "${CAL_DISTRO_DIR}"
+    rm -rf "${CAL_DISTRO_DIR}"
+    mkdir -p "${CAL_DISTRO_DIR}"
 
-set +o errexit
+    set +o errexit
 
-xcodebuild archive -workspace "${WORKSPACE}" -scheme "${SCHEME}" \
-    -configuration "${CONFIG}" -archivePath "${ARCHIVE_BUNDLE}" \
-    -sdk iphoneos | bundle exec xcpretty -c
+    xcodebuild archive -workspace "${WORKSPACE}" -scheme "${SCHEME}" \
+        -configuration "${CONFIG}" -archivePath "${ARCHIVE_BUNDLE}" \
+        -sdk iphoneos | bundle exec xcpretty -c
 
-RETVAL=${PIPESTATUS[0]}
+    RETVAL=${PIPESTATUS[0]}
 
-set -o errexit
+    set -o errexit
 
-if [ $RETVAL != 0 ]; then
-    echo "FAIL:  archive failed"
-    exit $RETVAL
+    if [ $RETVAL != 0 ]; then
+        echo "FAIL:  archive failed"
+        exit $RETVAL
+    fi
+
+    set +o errexit
+
+    xcrun -sdk iphoneos PackageApplication -v "${PWD}/${APP_BUNDLE_PATH}" -o "${PWD}/${IPA_PATH}" > /dev/null
+
+    RETVAL=$?
+
+    set -o errexit
+
+    if [ $RETVAL != 0 ]; then
+        echo "FAIL:  export archive failed"
+        exit $RETVAL
+    fi
+
+    cp "${IPA_PATH}" "${XAMARIN_DIR}/"
 fi
-
-set +o errexit
-
-xcrun -sdk iphoneos PackageApplication -v "${PWD}/${APP_BUNDLE_PATH}" -o "${PWD}/${IPA_PATH}" > /dev/null
-
-RETVAL=$?
-
-set -o errexit
-
-if [ $RETVAL != 0 ]; then
-    echo "FAIL:  export archive failed"
-    exit $RETVAL
-fi
-
-cp "${IPA_PATH}" "${XAMARIN_DIR}/"
